# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Media Server Orchestration API
  description: |
    Complete API for managing media server infrastructure including Docker services,
    configuration management, health monitoring, and seedbox integration.
  version: 1.0.0
  contact:
    name: Media Server API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3002
    description: Development server
  - url: https://api.mediaserver.local
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/services/start:
    post:
      summary: Start services
      description: Start one or more Docker services
      tags:
        - Services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceActionRequest'
      responses:
        '200':
          description: Services started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceActionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/services/stop:
    post:
      summary: Stop services
      description: Stop one or more Docker services
      tags:
        - Services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceActionRequest'
      responses:
        '200':
          description: Services stopped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceActionResponse'
  /api/services/restart:
    post:
      summary: Restart services
      description: Restart one or more Docker services
      tags:
        - Services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceActionRequest'
      responses:
        '200':
          description: Services restarted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceActionResponse'
  /api/services/{serviceName}/logs:
    get:
      summary: Get service logs
      description: Returns logs for a specific service
      tags:
        - Services
      parameters:
        - name: serviceName
          in: path
          required: true
          schema:
            type: string
        - name: lines
          in: query
          schema:
            type: integer
            default: 100
          description: Number of log lines to return
        - name: follow
          in: query
          schema:
            type: boolean
            default: false
          description: Follow log output
      responses:
        '200':
          description: Service logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'

  # Configuration Management Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: string
            details:
              type: string
            errorId:
              type: string
          required:
            - error

    # Health Schemas
...
    ServiceActionRequest:
      type: object
      properties:
        services:
          type: array
          items:
            type: string
          description: List of service names to act upon (empty for all services)
        profile:
          type: string
          enum: [minimal, media, download, monitoring, full]
          description: Service profile to use
    ServiceActionResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                services:
                  type: array
                  items:
                    type: string
                profile:
                  type: string
                stdout:
                  type: string
                stderr:
                  type: string

    # Configuration Schemas
...
    LogsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                logs:
                  type: array
                  items:
                    $ref: '#/components/schemas/LogEntry'
    LogEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [ERROR, WARN, INFO, DEBUG, TRACE]
        message:
          type: string
        meta:
          type: object
        service:
          type: string
      required:
        - timestamp
        - level
        - message
...